// src/modules/trending/models/trending.model.ts
import { ApiContextEnum } from '#common/enums'
import type { z } from 'zod'

export interface TrendingSongResponse {
  id: string;
  title: string;
  subtitle: string;
  header_desc: string;
  type: string;
  perma_url: string;
  image: string;
  language: string;
  year: string;
  play_count: string;
  explicit_content: string;
  list_count: string;
  list_type: string;
  list: string;
  more_info: {
    music: string;
    album_id: string;
    album: string;
    label: string;
    origin: string;
    is_dolby_content: boolean;
    '320kbps': string;
    encrypted_media_url: string;
    encrypted_cache_url: string;
    album_url: string;
    duration: string;
    rights: {
      code: string;
      cacheable: string;
      delete_cached_object: string;
      reason: string;
    };
    cache_state: string;
    has_lyrics: string;
    lyrics_snippet: string;
    starred: string;
    copyright_text: string;
    artistMap: {
      primary_artists: Array<{
        id: string;
        name: string;
        role: string;
        image: string;
        type: string;
        perma_url: string;
      }>;
      featured_artists: Array<any>;
      artists: Array<{
        id: string;
        name: string;
        role: string;
        image: string;
        type: string;
        perma_url: string;
      }>;
    };
    release_date: string;
    vcode?: string;
    vlink?: string;
    webp?: string;
    lyrics_id?: string;
  };
}

export interface TrendingSongsParams {
  language?: string;
  context?: ApiContextEnum;
}

export interface TrendingSongsResponse {
  status: string;
  message: string;
  data: TrendingSongResponse[];
}